+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2.5 03/07/2025
+### Adicionado
+- Relatório de Planos de Trabalho
+- Relatório de Planos de Entrega
+- Adição da Data de Recurso
+
 ## 2.4.2 03/07/2025
 ### Corrigido
  - Correção na edição de planos de entrega
@@ -8,6 +14,7 @@
 ### Corrigido
  - Bug na habilitação do pedágio
  - Bug na criação de plano de trabalho pelo perfil colaborador
+
 ## 2.4.0 11/06/2025
 ### Adicionado
  - Função para bloquear o teletrabalho para servidores em pedágio